+++ lib/kino/qx/application.ex
@@ -5,7 +5,7 @@ defmodule Kino.Qx.Application do
   @impl true
   def start(_type, _args) do
     Kino.SmartCell.register(Kino.Qx.SmartCell)
-    Kino.SmartCell.register(Kino.Qx.TranspileCell)
+    Kino.SmartCell.register(Kino.Qx.CredentialsCell)
     Supervisor.start_link([], strategy: :one_for_one, name: Kino.Qx.Supervisor)
   end
 end